Optimize Date.toJulianDate()
The original implementation of this function was using a specific algorithm, checked against a known data set.

Also add to the data set a value in the year 3001 for extra checking.

The 30ms max drift was not calculated, and it turns out that it can be higher. However, since the precision is 0.000001 of a Julian Day in our data set, we can look at +/- 1/2 of that as our zone of uncertainty, and our calculations should agree with the data set to that level. Turns out the minimum precision is about 43.2ms, which is good enough for this library.