Provide Quick Calculation for Sun/Moon Position
The result of these calculations are in RA/Dec, and would need to be converted to horizon coordinates if desired.

Not entirely sure how accurate this is just yet, but the solar appears to be within NREL’s calculator by about 0.38 degrees. Lunar is harder to test because there’s no real good data source for this.