Implement Sunrise / Moonrise Calculations
This is a somewhat large body of work that makes it possible to get rise and set times for the Sun and Moon.

Of course it includes full tests for Moonrise, Sunrise, Civil Twilight, Nautical Twilight, and Astronomical Twilight.

Also included is a refactor that pulls out some of the code from Sun/Moon so that it can be reused. The rise/set calculations is also on this refactoring, which is why we got Sun and Moon at the same time. Since I wrote it for the moon, I added tests for Sunrise/Sunset to confirm the results are also expected.

Accuracy is within a minute of https://www.timeanddate.com/sun/usa/seattle, in many cases much better than that, as the service rounds to the nearest minute, while my tests simply truncate.